+++ fhir-persistence-jdbc/src/main/java/com/ibm/fhir/persistence/jdbc/dao/api/IResourceReferenceDAO.java
@@ -86,5 +86,5 @@ public interface IResourceReferenceDAO {
      * @param canonicalValue
      * @return
      */
-    Integer readCanonicalId(String canonicalValue);
+    Long readCanonicalId(String canonicalValue);
 }
